Auto-fill date fields
Surround the date event and time entry duplication in a timeout. We
need to wait for the select boxes to be filled in before we can
duplicate the time entries, so wait a conservative amount of time to
ensure the first entry is filled in before duplicating it.

The date event listener is also part of this timeout as it fires an
event on load without user interaction. We don't care about that
initial event.

Our date listener will listen for changes to the first time entry's
date field and automatically fill in the four duplicated entries with
subsequent dates increasing by 1 for each entry. This allows us to
select a Monday for the first time entry and have the rest of the week
automatically filled in for us.