timetasker.js: Load fields from options
Query the Chrome storage API to get user options for time entry
fields. Replace our hard-coded values with user-entered ones.

This allows the extension to work for others without requiring
modification of the source code.

In order to be able to pull the options, I had to move the main
extension code into a callback function that can be passed to
`chrome.storage.sync.get`.