timetasker.js: Fix the "billable" checkbox auto-fill
The checkbox wasn't getting filled in correctly. By some overzealous and
uncaught copy-pasting, I had wrapped the field in the function to
populate a select box. Since this is a checkbox, we want to use its
`.checked` property instead.

Use a timeout also. By default Time Task marks Billable as true/checked.
If the script unchecks the box and then the AJAX select box population
occurs, the check box re-checks itself. Adding a timeout allows it to
remain unchecked if you've set that in the user options.